+++ app/packs/src/components/navigation/Navigation.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Nav, Navbar, Tooltip, OverlayTrigger } from 'react-bootstrap';
 import UserAuth from 'src/components/navigation/UserAuth';
-import Search from 'src/components/search/Search';
+import Search from 'src/components/navigation/search/Search';
 import ManagingActions from 'src/components/managingActions/ManagingActions';
 import ContextActions from 'src/components/contextActions/ContextActions';
 import UserStore from 'src/stores/alt/stores/UserStore';
@@ -13,7 +13,7 @@ import NavNewSession from 'src/components/navigation/NavNewSession'
 import NavHead from 'src/components/navigation/NavHead'
 import DocumentHelper from 'src/utilities/DocumentHelper';
 import NavigationModal from 'src/components/navigation/NavigationModal';
-import SearchFilter from 'src/components/search/SearchFilter.js'
+import SearchFilter from 'src/components/navigation/search/SearchFilter.js'
 
 const colMenuTooltip = <Tooltip id="col_menu_tooltip">Toggle sidebar</Tooltip>;
